+++ Classes/NSInvocationAdditions.m
@@ -51,7 +51,7 @@ THE POSSIBILITY OF SUCH DAMAGE.
 {
     id retval=nil;
     [self invokeWithTarget:target];
-    int returnLength=[[self methodSignature] methodReturnLength];
+    long returnLength=[[self methodSignature] methodReturnLength];
     if (returnLength > 0 && returnLength <= sizeof retval) {
         [self getReturnValue:&retval];
     }
